Make LegacyLabware take Labware object
